feat: change code distribution
