+++ 04-ciencia-da-computacao/m04-section-06-data-structure-ii/m04s06d01-hashmap-and-dict/m04s06d01-follow-material/employee.py
@@ -1,24 +1,20 @@
+import pytest
+
+
 class Employee:
     def __init__(self, id_num, name):
         self.id_num = id_num
         self.name = name
 
 
-class HashMap:
-    def __init__(self):
-        self._buckets = [None for _ in range(10)]
+def test_employee_create():
+    employee = Employee(14, "name2")
 
-    def get_address(self, id_num):
-        return id_num % 10
+    assert isinstance(employee, Employee)
 
-    def insert(self, employee):
-        address = self.get_address(employee.id_num)
-        self._buckets[address] = employee
-
-    def get_value(self, id_num):
-        address = self.get_address(id_num)
-        return self._buckets[address].name
-
-    def has(self, id_num):
-        address = self.get_address(id_num)
-        return self._buckets[address] is not None
+    with pytest.raises(TypeError):
+        Employee()
+    with pytest.raises(TypeError):
+        Employee(1)
+    with pytest.raises(TypeError):
+        Employee("name")
